Use #prepend for decorators
